+++ src/diag.c
@@ -409,6 +409,7 @@ static const struct {
    { "Qualified expressions", { [STD_08] = "9.3.5" } },
    { "Interface package declarations", { [STD_08] = "6.5.5" } },
    { "Selected names", { [STD_08] = "8.3", [STD_93] = "6.3" } },
+   { "Simple variable assignments", { [STD_08] = "10.6.2" } },
 };
 
 diag_t *diag_new(diag_level_t level, const loc_t *loc)